Add a distributed logic to clean up shared memory and improved code structure in general

+++ streaming/base/util.py
@@ -4,18 +4,18 @@
 """Utility and helper functions for datasets."""
 
 import os
-
 from multiprocessing.shared_memory import SharedMemory as BuiltinSharedMemory
 from time import sleep, time
 from typing import List
 
-from streaming.base.shared._constant import (BARRIER, CACHE_USAGE, EPOCH_DATA, EPOCH_SHAPE, LOCALS,
+import torch.distributed as dist
+
+from streaming.base.distributed import get_local_rank, maybe_init_dist
+from streaming.base._constant import (BARRIER, CACHE_USAGE, EPOCH_DATA, EPOCH_SHAPE, LOCALS,
                                              NEXT_EPOCH, RESUME, SHARD_ACCESS_TIMES, SHARD_STATES)
+from streaming.base.shared.prefix import _create_filename
 
 __all__ = ['get_list_arg']
-
-# Time to wait, in seconds.
-TICK = 0.007
 
 
 def get_list_arg(text: str) -> List[str]:
@@ -96,24 +96,40 @@ def bytes_to_int(bytes_str: str) -> int:
 
 
 def clean_stale_shared_memory() -> None:
-    """Clean up all the leaked shared memory."""
-    shm_names = [
-        BARRIER, CACHE_USAGE, EPOCH_DATA, EPOCH_SHAPE, LOCALS, NEXT_EPOCH, RESUME,
-        SHARD_ACCESS_TIMES, SHARD_STATES
-    ]
-    for prefix_int in range(1000000):
-        leaked_shm = False
-        prefix = f'{prefix_int:06}'
-        for shm_name in shm_names:
-            name = f'{prefix}{shm_name}'
-            try:
-                shm = BuiltinSharedMemory(name, True, 4)
-            except FileExistsError:
-                shm = BuiltinSharedMemory(name, False, 4)
-                leaked_shm = True
-            finally:
-                shm.close()  # pyright: ignore
-                shm.unlink()
-        # Break if no leaked shared memory
-        if not leaked_shm:
-            break
+    """Clean up all the leaked shared memory.
+
+    In case of a distributed run, clean up happens on local rank 0 while other local ranks wait for
+    the local rank 0 to finish.
+    """
+    # Initialize torch.distributed ourselves, if necessary.
+    destroy_dist = maybe_init_dist()
+
+    # Perform clean up on local rank 0
+    if get_local_rank() == 0:
+        shm_names = [
+            BARRIER, CACHE_USAGE, EPOCH_DATA, EPOCH_SHAPE, LOCALS, NEXT_EPOCH, RESUME,
+            SHARD_ACCESS_TIMES, SHARD_STATES
+        ]
+        for prefix_int in range(1000000):
+            leaked_shm = False
+            for shm_name in shm_names:
+                name = _create_filename(prefix_int, shm_name)
+                try:
+                    shm = BuiltinSharedMemory(name, True, 4)
+                except FileExistsError:
+                    shm = BuiltinSharedMemory(name, False, 4)
+                    leaked_shm = True
+                finally:
+                    shm.close()  # pyright: ignore
+                    shm.unlink()
+            # Come out of loop if no leaked shared memory
+            if not leaked_shm:
+                break
+
+    # Sync all ranks
+    if dist.is_available() and dist.is_initialized():
+        dist.barrier()
+
+    # Delete the process group if Streaming initialized it.
+    if destroy_dist:
+        dist.destroy_process_group()
